WIP - Restructured code
